Fold the five single-holiday commands into the /event group
The bot sat at 98 of Discord's 100 global command slots. That limit is not
a soft one: the deploy is a single PUT and Discord rejects the whole
payload past 100 rather than registering the first hundred, so one command
too many unregisters every other command — in production, at deploy time,
with no local signal before it. commandDeployer's guard fails the deploy
at 101 instead, which is right, but it left two slots for every future
feature.

Five of the ninety-eight were single-holiday activities usable a few weeks
a year — /snowball, /sandcastle, /trickortreat, /lovenote, /trackhunt —
and every one of them already opened by asking seasonalEventService
whether its event was live. They are subcommands of /event now, the
command that starts and ends those events, which is where a player looking
for 'what can I do during this event' would look first anyway. Seven
top-level commands become one and COMMAND_BUDGET drops 97 -> 92.

/event held the name already, as commands/admin/event.js, so the two
halves merged rather than one displacing the other. It moves to
economy/, because six of its eight subcommands are things players do and
only start and end are administrative. Those two keep the inline Manage
Server check they already had — Discord's setDefaultMemberPermissions
gates a whole command, so using it here would hide the five activities
from everyone without Manage Server.

Two things change beyond the move. Each subcommand now gets its own
process-local cooldown bucket via cooldownKey, where five separate
commands had five separate buckets before and a naive fold would have made
a snowball throw eat the sandcastle window; the activities pass 0 because
their real cooldowns are claimed atomically in Mongo, as they always were.
And /event lovenote joins the cooldown-claim guard's table, which had
listed its four siblings and not it.

CONTRIBUTING.md now says what the budget means and that a new feature
ships as a subcommand of an existing group unless there is a reason it
cannot — the ratchet in tests/commandCap fails when the count moves, but
it could not say what to do about it.

Closes #875
