+++ docs/COMMANDS.md
@@ -12,7 +12,7 @@ commands, not this file.
 
 <!-- BEGIN GENERATED COMMANDS — npm run docs:commands -->
 
-_Generated by `npm run docs:commands` from the loaded command set — 98 entries. Edit the commands, not this list._
+_Generated by `npm run docs:commands` from the loaded command set — 93 entries. Edit the commands, not this list._
 
 ### 📦 Economy
 
@@ -25,6 +25,7 @@ _Generated by `npm run docs:commands` from the loaded command set — 98 entries
 - `/daily` — Claim your daily coin reward (amount set by server admins, default 100). Resets every 24 hours.
 - `/dailychallenge` — View and claim your Prestige VI+ Daily Challenge board bonus.
 - `/duel` — Challenge another user to a coin-bet duel
+- `/event` — Limited-time events: play the current one, or manage it.
 - `/eventshop` — Browse and purchase items with your event currency
 - `/explore` — World exploration: set out into the wilds, chart your map, and find what hides there.
 - `/featured` — Today's featured rotation — +25% payout and +10% rare chance on each highlighted option.
@@ -36,7 +37,6 @@ _Generated by `npm run docs:commands` from the loaded command set — 98 entries
 - `/inventory` — View your or another user's inventory
 - `/invest` — Contribute coins to server districts and unlock server-wide benefits.
 - `/jobs` — Browse all available jobs, their tiers, and pay ranges
-- `/lovenote` — Send a love note into the Arcade and see what comes back! Valentine's Day only.
 - `/market` — Server player-to-player item marketplace.
 - `/mine` — Mining: dig, profile, inventory, quests, and shop.
 - `/pet` — Manage your pets.
@@ -44,16 +44,12 @@ _Generated by `npm run docs:commands` from the loaded command set — 98 entries
 - `/quiz` — Answer a trivia question to win coins — or lose some if you're wrong!
 - `/rob` — Try to rob another member's wallet. Success is affected by tools and protection.
 - `/robstatus` — Spy on a target's active rob protections before committing to a heist.
-- `/sandcastle` — Build a sandcastle on the shore! Summer Festival only.
 - `/season` — Season pass and economy season commands.
 - `/shop` — Browse and buy items from the server shop
 - `/showcase` — Display a player's trophy card — rarest items and top achievements.
-- `/snowball` — Throw a snowball at another user! Only available during Winter Wonderland.
 - `/syndicate` — Crime syndicate system — cooperative heists, territory, and rivalry.
 - `/synergies` — View all cross-system synergy bonuses and your progress toward unlocking them
-- `/trackhunt` — Track game across the Arctic Tundra! The Winter Hunt only.
 - `/trap` — Set a hidden tripwire on your wallet. Triggers if someone successfully robs you.
-- `/trickortreat` — Knock on a spooky door for candy or consequences! Spooky Season only.
 - `/use` — Use an item from your inventory
 - `/war` — Server vs server war events.
 - `/winfeed` — See the last 20 big wins in this server (50k+ coins or legendary drops).
@@ -133,7 +129,6 @@ _Generated by `npm run docs:commands` from the loaded command set — 98 entries
 
 ### ⚙️ Admin
 
-- `/event` — Manage limited-time events
 - `/raidmode` — Configure raid detection and case management settings
 
 <!-- END GENERATED COMMANDS -->
